refactor(ros2): upgrade to ros rolling ridley
which we will be using for the bit-bots codebase ros2 migration.
It is to be noted that the following packages are not available for
the rolling release:
- ros-rolling-joint-state-controller
- ros-rolling-navigation2
- ros-rolling-robot-controllers

While the joint-state-controller seems to have moved to the packages
ros-rolling-joint-state-broadcaster, which is installed instead now both
navigation2 as well as the robot-controllers are not available at all
yet for the rolling release.
For this reason they have been added as submodules in the bitbots_meta
repo for manual building bit-bots/bitbots_meta@97c10a5.